chore(analyzer): move ffmpeg and utils modules to pipeline
