+++ analyzer/libretime_analyzer/pipeline/_ffmpeg.py
@@ -3,7 +3,7 @@ from math import inf
 from pathlib import Path
 from typing import List, Optional, Tuple
 
-from .utils import run_
+from ._utils import run_
 
 
 def _ffmpeg(*args, **kwargs):
